Refatorando mais a arquitetura
