+++ src/paginas/dres/DadosDaAssociacao/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+import {PaginasContainer} from "../../PaginasContainer";
+import {DadosDaAssociacao} from "../../../componentes/dres/Associacoes/DadosDasAssociacoes/DadosDaAssociacao";
+
+export const DadosDaAssociacaoDrePage = () =>{
+    return (
+        <PaginasContainer>
+            <DadosDaAssociacao/>
+        </PaginasContainer>
+    )
+};
