+++ src/paginas/dres/DadosDaUnidadeEducacional/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import {PaginasContainer} from "../../PaginasContainer";
 import {DadosDaUnidadeEducacional} from "../../../componentes/dres/Associacoes/DadosDasAssociacoes/DadosDaUnidadeEducacional";
 
-export const DadosDasAssociacoesPage = () =>{
+export const DadosDaUnidadeEducacionalPage = () =>{
     return (
         <PaginasContainer>
             <DadosDaUnidadeEducacional/>
